requirements.txt
anthropic
gradio
duckduckgo-search
requests
python-dotenv
